+++ api/src/main/java/infidelity/api/data/Coin.java
@@ -1,10 +1,14 @@
 package infidelity.api.data;
 
+import lombok.Getter;
+
 import javax.persistence.Column;
+import javax.persistence.Entity;
 import javax.persistence.Id;
 
+@Entity
+@Getter
 public class Coin extends Tradeable{
-    @Id
     @Column(nullable = false)
     private String name;
 }
